Unfinished Frontend Mentor Challenege github user search
Started this challenge back in late 2022, got the JS portion working fine but fell away from it. Did a little work on it recently but submitting to get it in the repo. Not yet in the projects list. The time tracking dashboard is intended as what I was going to attempt next (soon)